+++ src/org/torproject/descriptor/BridgeServerDescriptor.java
@@ -1,0 +1,9 @@
+/* Copyright 2015 The Tor Project
+ * See LICENSE for licensing information */
+package org.torproject.descriptor;
+
+/* Contains a bridge server descriptor. */
+public interface BridgeServerDescriptor extends ServerDescriptor {
+
+}
+