+++ src/org/torproject/descriptor/RelayExtraInfoDescriptor.java
@@ -1,0 +1,9 @@
+/* Copyright 2015 The Tor Project
+ * See LICENSE for licensing information */
+package org.torproject.descriptor;
+
+/* Contains a relay extra-info descriptor. */
+public interface RelayExtraInfoDescriptor extends ExtraInfoDescriptor {
+
+}
+